github prci: periodically reboot runner after 3 hours
When the runner is online for a long time, the virtualized guests
eventually slow down to a crawl, rendering them unusable. Investigation
revealed CPU slowdown by hardware interrupts (40-60%) and very slow
system calls to get time.

Switching to other clocksources than kvm-clock in the guest doesn't seem
to solve the issue. Restarting the libvritd service on host also doesn't
help. The host's clock or performance doesn't seem to be affected.
Rebooting the machine seems to help the issue.

This is probably some complex bug caused by the nested virtualization
and spinning up too many machines and throwing them away afterwards.

Since the cost of reboot is fairly low (just some github API calls), it
seems to be a usable workaround for the time being.

Signed-off-by: Tomas Krizek <tkrizek@redhat.com>