github: Recovery from runner craches
When runner takes task and then craches the task would become stale and manual
re-run would be required. Since time when the task in included in task
description and timeout for the task is specified in task configuration we can
detect such situation and regenerate such task so other runner can take it.