Allowing multiple topologies
In order to have more tests running in upstream PR CI
infrastructure, we need to support multiple topologies:
multiple master, replicas and clients. This PR address this.

To understand the resources that a given task should have to
run it, the freeipa/freeipa/.freeipa-pr-ci.yml file will
declare how much RAM and CPUs is needed. Using this syntax:

```
topologies:
  master_1repl: &master_1repl
    name: master_1repl
    cpu: 2
    memory: 4098
  master_1repl_1client: &master_1repl_1client
    name: master_1repl_1client
    cpu: 6
    memory: 6144
```

After declaring the topologies, the tasks will be able
to use it with this syntax:

```
fedora-26/simple_replication:
    ...
    job:
      class: RunPytest
      args:
        ...
        topology: *master_1repl
```

Is IMPORTANT to say that, the name of the topology
must match with the name of a Vagrantfile inside the
templates/vagrantfiles dir. So, if a topology is added to
the .freeipa-pr-ci.yml file, a correspoding Vagrantfile
must be added in the freeipa-pr-ci project as well.

The Task class will read the config file and the Scheduler
will deal with it, to know if the runner is able to run
that task.

Fixes #70