Adding installed packages list to a separate file
With this change the installed packages will be listed in the
/vagrant/installed_packages/installed_packages_<hostname>.log.gz
file. After sending the stdout to the file, we have a task that
collects the log files in the shared directory. This is done via the
shared directory inside the vagrant vms in /vagrant.
For more info about it, check tasks/remote_storage.py::GzipLogFiles

Fixes #106