Adding the tox tests on PR CI infraestructure
A simple Vagranfile was created to be used to run the tox tests.
A new class ToxTests was added and can be used in .freeipa-pr-ci.yaml
in freeipa. Also, news packages were added in the creation of the
template in order to have they available to run the tox tests.

Fixes #42