Enabling multiple topologies
In order to have more tests running in upstream PR CI
infrastructure, we need to support multiple topologies:
multiple master, replicas and clients. This PR address this.

To know the resources that a given task should have to
run it, the freeipa/freeipa/.freeipa-pr-ci.yml file will
declare how much RAM and CPUs are needed. Using this syntax:

```
topologies:
  master_1repl: &master_1repl
    name: master_1repl
    cpu: 2
    memory: 3800
  master_1repl_1client: &master_1repl_1client
    name: master_1repl_1client
    cpu: 4
    memory: 6400
```

After declaring the topologies, the tasks will be able
to use it with this syntax:

```
fedora-26/simple_replication:
    ...
    job:
      class: RunPytest
      args:
        ...
        topology: *master_1repl
```

It's IMPORTANT to say that, the name of the topology
must match with the name of a Vagrantfile inside the
templates/vagrantfiles dir. So, if a topology is added to
the .freeipa-pr-ci.yml file, a corresponding Vagrantfile
must be added to the freeipa-pr-ci project as well.

When a runner tries to execute a task that it doesn't the
capacity, the runner drop the task and it changes the
task status on GitHub back to unassigned.

Fixes #70