CLI tool to open/close pull requests
A new CLI tool to open and close Pull Requests. The tool
receives the arguments:
- identifier: that will be used to control identify the
  branch
- branch: the branch that the PR will be open against it
- repo_owner: the owner of the freeipa repo
- github_token: to use the API
- prci_config: a custom .freeipa-pr-ci.yaml file to change
  the default one
- git_sshkey: a ssh key to pull and push to the fork

Before opening a new PR, the tool create a branch using the
the base branch from the upstream repo.

This also creates an Ansible script to configure this
tool in an existing runner. You can use it like this:
```
ansible-playbook -i ansible/hosts/runners \
        ansible/prepare_openclose_pr_tool.yml
```

Fixes: #109