Increase the time between requests
As commit a811f1cc explains, we need to slow down the API calls to avoid
hitting the limit of requests per minute. To have more than ~20
runners, it's necessary to increase this time.
Increasing it to 1.5s would allow us to have around 40 runners.

Important to say that the limit is hit even when runners are using
the cached requests, as it's well explained in [1].

[1] https://github.com/freeipa/freeipa-pr-ci/tree/master/doc#abuse-limit