Do not use API call for downloading tasks file
Now downloading tasks file is using an GitHub API call
(GET /repos/:owner/:repo/pulls/:number/files). It can be useful
in some cases but not in ours. In our case the `.freeipa-pr-ci.yaml`
file exists either in the head or base branches. So we can retrieve
it just by constructing the URL. So then if the tasks file doesn't
exist in the head, we will just download it from the base.