Use GraphQL GitHub API alonside with REST
After we added more tests for PR to perform on CI, we're hitting
the GitHub REST API rate limit and overall performance of CI is
dramatically low.

Also, we have a problem that code which responsible for communication
with GitHub API is really hard to understand.

This patch is intended to fix these issues.

As GitHub also has GraphQL API endpoint which allows us to get all
the required information about every pull request in one API call
and its rate limit is different from REST API rate limit, this
patch is implementing the way of communicating with GitHub using
two GitHub API endpoints.

All the information gathering (think of it as GET requests) is
happening using GraphQL API. All the mutations of a state of pull
requests and commits statuses are happening through REST API, as
for now GraphQL API has really limited mutations support (think of it
as POST and PUT).

To achieve this way of APIs usage it was necessary to replace
one part of PR CI completely as it was intended to heavely use
github3py library in many places in the code.

Now there is World class which is responsible for communication
with outside world to change its state and a bunch of data classes
to map the data recieved from GitHub.

Also, I've added type hintings for a changed code so it will be
more easily for other contributors to understand this part of CI.

And in the end, I've wrote a bit of sanity unit tests for data
classes.

Auto update and auto restart of runners were intentionally removed
as for auto update it's bad in the first place to update infrastructure
without supervising and as for restart, nobody gave me the real reason
why it was implemented in the first place.