Use tuned to make optimizations for runners
Tuned is a profile-based system tuning tool that uses the udev device
manager to monitor connected devices, and enables both static and
dynamic tuning of system settings.

As the runner is virtual hosts for Vagrant machines it's good to have
it's kernel optimized for such workloads. To achieve this we need to
have tuned service enabled and running. Then, we need to activate a
virtual-host profile so tuned will make necessary changes in sysctl.

>> virtual-host decreases the swappiness of virtual memory and enables
>> more aggressive writeback of dirty pages.