[scalability] Fix behaviour to update stalled task
Now prci will try to set rerun pending for stalled task and this
update will fail as set_rerun has check if task is taken. In this
case the update of a stalled task will be performed over and over
again that will lead to triggering of abuse limit of GitHub API.

This patch makes the prci to set stalled task status to unassigned
and fixes this problem.