Fix failure after PR is forced pushed
When PR is forced pushed when test is running. The test will be
finished and then PR-CI will error out on status update step with

ERROR: __main__: Can't parse status data.

It is expected as the force push removed the status. But it
causes upload of data to sentry.io.

This patch adds EnvironmentError exception catching into execute
method of a Task class and raises ReferenceError when commit status
is not found. The error logs at a warning level, then runner's
execution proceeds.