Making PR CI nightly tests definition editable by FreeIPA contributors
Now the .freeipa-pr-ci.yaml it's just a symlink to a file that actually
has the definition of the tests. This commit changes the tool that
schedules the nightly tests.

The major change is that the "--prci_config" parameter now receives a
relative path that will change to where the .freeipa-pr-ci.yaml link points