Changing volume caching from 'none' to 'unsafe'
Qcow2 file backend can be significantly slow with cache mode 'none'
setting. This happens because the actual file data is fragmented. In
other words, the qcow2 file remain a sparse one with only short stroke
of allocation (for metadata).

So it's better to have an aggressive VFS RAM caching to have a better
random access speed.

The key aspect of this unsafe mode, is that all flush commands from the
guests are ignored. Using this mode implies that the user has accepted
the trade-off of performance over risk of data loss in the event of a
host failure. As we are using virtual machines only for a short period
of time it's safe to use it.