introduce runner autocleaner
It happens that we hit an "out of disk space" issue on one of
our runners. This usually happens due to dangling libvirt images
which were left behind when VM erase was not completed completely.

Also for future automatic template generation we need to be able
to get rid of old template versions.

Autocleaner is doing the following:

1. Cleaning Vagrant boxes which are not used in any active branch

- go through current Vagrant boxes on particular runner
- get branch from box name
- perform GH GraphQL query to get PRCI yaml definitions files
- check if the box is defined anywhere
- if not, the box is deleted together with libvirt image

2. Clean dangling libvirt images when VM erase was not completed
   completely.

- find old VM images in libvirt directory by UUID and delete them

3. If "--jobs_dir_exp <DAYS>" argument is provided, delete PRCI
   job directories which are older then <DAYS>

Before performing the cleaning task, autorunner checks if there is
any qemu process running to see if PRCI is active and if not process
is found it stops PRCI service and performs the task. The service is
then again started.