Install, configure and mask firewalld in the machine setup
Install firewalld in the machine setup, enable ssh firewall service and
mask firewalld.

For now firewalld gets masked after it has been configured to make sure
that tests that are not able to configure firewalld are still usable.

firewalld is configured using the command line tool. This has been done to
make sure that there is no Python2/3 import issue for the ansible firewalld
module is the firewalld imports are not available for this Python version.
The use of the command line tool has an advantage though: It is only one
call to enable all needed services at once instead of one call per service.

The firewalld service is masked. The mask should to be replaced with a
firewalld enable and start if all tests are able to configure the firewall.

The reload and also the configuration using the firewalld ansible module
are added to the sections, but commented out.

This is needed for:
https://pagure.io/freeipa/issue/7755 (Enable firewall in the tests)
https://github.com/freeipa/freeipa/pull/2542 (Enable firewall in the tests)