Remove decprecated join(' ') filter from machine/setup role
Due to some updates in various modules, they start to prefer to treat
a list instead of doing a loop or call a function.

The usage of join(' '), in fedora 29 ansible version, was breaking the
placeholder interpolation and resulting in execution failure.

Since 'python_packages_to_install' placeholder is already defined as a
list, we don't need the join(' ') anymore, so this commit is removing it.

The documentation of ansible pip module:

https://docs.ansible.com/ansible/2.7/modules/pip_module.html

Signed-off-by: Diogo Nunes <diogo.hap@gmail.com>