scheduler_tool: define prci_config as part of nightly jobs definition
"prci_definition" defines the set of tests to be executed. It is different for each
nightly job, hence this commit defines the proper prci_definition per nightly flow.
Same happens for "identifier" variable so this commit removes this variable prompt as
well. In addition, openclose_pr folder is deleted as it is no longer used, and the
documentation has been updated accordingly.