when rebooting windows wait for successful operation on AD
dsqury was used to check if Windows and AD service is up. But there is
a small delay between start of ldap server and AD service being actually
available which caused errors when trying to create test objects in AD.
Now we use command Get-ADUser which requires AD to be fully functional.