Increase task extra time to avoid job jumping to another runner
Runners are considering jobs that are still being executed
by another runner as stalled. Some of those jobs are about to time-out;
however, a new runner marks it as stalled and re-runs it, while the
first runner is still running the job. Github doesn't receive an
error/failure from the first runner, only the notification that a
new runner is picking the job.

Issue: https://github.com/freeipa/freeipa-pr-ci/issues/365

Signed-off-by: Armando Neto <abiagion@redhat.com>