Make lock mechanism quicker and clearer
By removing the sleep when locking a task the time between polling the
status and setting a new one is reduced, this is important to lower the
chances of a race condition.

"Taken" is being divided into "Locked" and "Taken", that can also help us
to troubleshoot future issues.

"pending for rerun" status now includes the runner that set it and when
it was set.

Issue: https://github.com/freeipa/freeipa-pr-ci/issues/365

Signed-off-by: Armando Neto <abiagion@redhat.com>