Disable bash bracketed-paste mode
enable-bracketed-paste is enabled by default since bash-5.1. When set to On,
readline will configure the terminal in a way that will enable it to insert
each paste into the editing buffer as a single string of characters, instead
of treating each character as if it had been read from the  keyboard. This is
affecting some of our automated tests:

pexpect.exceptions.ExceptionPexpect: Program exited with an unexpected state:
Unexpected output at program exit: '  \x1b[?2004l\r'

This commit adds bash_settings task group as part of utils role so that
the setting is system-wide disabled back.