fix(autocleaner): featching of prci def files from new branches
the autocleaner was failing on fetching prci def files from branch
ipa-4-10.  This happened because the files were downloaded from
PR-CI user's freeipa fork and not from the main "freeipa" organization.

This happened as the query was starting with 'viewer' query, which
represents the currently logged-in user.

This patch is changing the query to always check the main
freeipa/freeipa repo and thus it won't fail if the token owner repo
doesn't contain some new branch.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>