fix: open_close_pr to re-try PR creation on error
Rarely the open_close_pr receives error from GitHub when opening
a PR with message, e.g.:

"""
A pull request already exists for freeipa-pr-ci2:testing_master_latest
"""

The assumtion is that it might be caused by a race condition between
deletion of branch of previous PR and GitHub not closing the previous PR
fast enough.

This fix adds a delay before opening the MR to try to lower the
probability of this race condition + it tries the opening up to 3 times.

The fix also moves out assignment of reviewer and deletion of local
branch out of opening of PR to more clearly separate concerns of the
individual methods.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>