fix: do not start PRCI in deployment
The playbooks are changed to configure PR-CI but to keep it in its
current state.

This fixes issues when PR-CI is started before fully deployed and this
it wan't start processing the jobs.

Not starting the service also gives better control to the deployment,
e.g. to fix issues on other runners before starting entire fleet.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>