Provision machines: fix hostname logic
When the test machines are provisioned, the hostname must
be set (otherwise the default hostname "localhost" is used.

The current logic relies on the inventory hostname being either
master / client / replica / controller / trustedmaster /
trustedclient / trustedreplica but depending on the
Vagrant file the inventory hostname can also be client0 replica2 etc...

In this case the hostname is not set.

Change the logic to set the hostname to xx.ipa.test if the
inventory_hostname does not contain trustedxx and to
xx.trustedipa.test if the inventory_hostname contains
trustedxx.

Signed-off-by: Florence Blanc-Renaud <flo@redhat.com>