Cygwin installation: define default value in the loop
The playbook installing cygwin openssh contains a loop
trying multiple mirrors until one of them succeeds. Ansible
changed behavior and the result.failed variable is not set
if the iteration succeeded.
Default to false if not set, so that the loop continues
if the last iteration failed but is skipped if it succeeded.

Resolves: https://github.com/freeipa/freeipa-pr-ci/issues/512

Signed-off-by: Florence Blanc-Renaud <flo@redhat.com>